Added simple multimap impl to allow multiple values for a given key.
Refactored to use case-insensitive multimap impl for QueryMap and HeaderMap.

The map preserves the format of the key used to add a value, and attempts
to avoid changing semantics of existing code by multimap-type operations
being separate methods. Use operations such as `getAllEntries` to get
a list of key-value pairs for a given key.

Also added tests