refactor(manager-api): adding service layer
Currently most of Apiman's business logic is in the presentation layer.
This is going to be split into service layers to improve code-reuse and allow better abstractions
over time to improve various issues with the persistence layer being tightly coupled.

Also introducing the use of a `@Transactional` annotation instead of manual TX everywhere.