fix(manager-api): PolicyDefinitionTemplate missing #equals and #hashCode
PolicyDefinitionTemplate is referenced in a Set from PolicyDefinitionBean.
The lack of #equals and #hashCode can cause Hibernate to duplicate the templates and write them back
to the the database under certain circumstances.

Once the bug is triggered, the templates will often double in size on each occasion, which causes
exponential growth.

Fixes: #1566